reverted from size_t to int

+++ src/add.c
@@ -33,7 +33,7 @@ static char *add_same_len(const char *s1, const char *s2, const int len, int *rl
 	ans[0] = ZERO; 
 	
 	// add pointwise
-	for (size_t i = 1; i <= len; i++) {
+	for (int i = 1; i <= len; i++) {
 		if (s1[i - 1] != s2[i - 1]) {
 			ans[i] = ONE;
 		} else if (s1[i - 1] == s2[i - 1] && s1[i - 1] == ONE) {
@@ -46,7 +46,7 @@ static char *add_same_len(const char *s1, const char *s2, const int len, int *rl
 	// first stage
 
 	// eliminate TWOs
-	for (size_t i = 0; i <= len - 2; i++) {
+	for (int i = 0; i <= len - 2; i++) {
 		if (i < len - 2 && ans[i] == ZERO && ans[i + 1] == TWO && ans[i + 2] == ZERO) {
 			ans[i] = ONE;
 			ans[i + 1] = ZERO;
@@ -93,7 +93,7 @@ static char *add_same_len(const char *s1, const char *s2, const int len, int *rl
 	// second stage
 	
 	// first pass from right to left
-	for (size_t i = len; i >= 2; i--) {
+	for (int i = len; i >= 2; i--) {
 		if (ans[i] == ONE && ans[i - 1] == ONE && ans[i - 2] == ZERO) {
 			ans[i] = ZERO;
 			ans[i - 1] = ZERO;
@@ -102,7 +102,7 @@ static char *add_same_len(const char *s1, const char *s2, const int len, int *rl
 	}
 
 	// second pass from left to right
-	for (size_t i = 0; i <= len - 2; i++) {
+	for (int i = 0; i <= len - 2; i++) {
 		if (ans[i] == ZERO && ans[i + 1] == ONE && ans[i + 2] == ONE) {
 			ans[i] = ONE;
 			ans[i + 1] = ZERO;
